+++ src/background.vala
@@ -1,6 +1,12 @@
 using Cairo;
 
 namespace LiveChart { 
+
+    public struct LinearGradient {
+        public Gdk.RGBA from;
+        public Gdk.RGBA to;
+    }
+
     public class Background : Drawable, Object {
         private BoundingBox bounding_box = BoundingBox() {
             x=0, 
@@ -20,6 +26,8 @@ namespace LiveChart {
             }
         }
 
+        public LinearGradient? gradient { get; set; }
+
         [Version (deprecated = true, deprecated_since = "1.8.0", replacement = "Background.color")]
         public Gdk.RGBA main_color { 
             get; set; default= Gdk.RGBA() {
@@ -34,8 +42,16 @@ namespace LiveChart {
             if (visible) {
                 this.update_bounding_box(config);
                 ctx.rectangle(0, 0, config.width, config.height);
-                ctx.set_source_rgba(color.red, color.green, color.blue, color.alpha);
-                ctx.fill();
+                if (gradient != null) {
+                    Cairo.Pattern pattern = new Cairo.Pattern.linear(config.width/2, 0, config.width/2, config.height);
+                    pattern.add_color_stop_rgba(0, gradient.from.red, gradient.from.green, gradient.from.blue, gradient.from.alpha);
+                    pattern.add_color_stop_rgba(1, gradient.to.red, gradient.to.green, gradient.to.blue, gradient.to.alpha);
+                    ctx.set_source(pattern);
+
+                } else {
+                    ctx.set_source_rgba(color.red, color.green, color.blue, color.alpha);
+                }
+                ctx.fill();                
             }
         }